Added review check for course cards
